requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.58.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.0
tqdm==4.67.1
twilio==9.3.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.16.0
pytest==8.3.4
pytest-asyncio==0.25.2